Fab update

+++ code/game/mecha/mech_fabricator.dm
@@ -11,6 +11,9 @@
 	req_access = list(access_robotics)
 	circuit = /obj/item/circuitboard/mechfab
 
+	var/list/href_data
+	var/topic_result
+
 	var/speed = 1
 	var/mat_efficiency = 1
 	materials = list(MAT_STEEL = 0, "glass" = 0, "plastic" = 0, MAT_GRAPHITE = 0, MAT_PLASTEEL = 0, "gold" = 0, "silver" = 0, MAT_LEAD = 0, "osmium" = 0, "diamond" = 0, MAT_DURASTEEL = 0, "phoron" = 0, "uranium" = 0, MAT_VERDANTIUM = 0, MAT_MORPHIUM = 0, MAT_METALHYDROGEN = 0, MAT_SUPERMATTER = 0)
@@ -109,6 +112,7 @@
 	if(current)
 		data["builtperc"] = round((progress / current.time) * 100)
 
+	href_data = data
 	ui = SSnanoui.try_update_ui(user, src, ui_key, ui, data, force_open)
 	if(!ui)
 		ui = new(user, src, ui_key, "mechfab.tmpl", "Exosuit Fabricator UI", 800, 600)
@@ -117,6 +121,7 @@
 		ui.set_auto_update(1)
 
 /obj/machinery/mecha_part_fabricator/Topic(href, href_list)
+	topic_result = href_list
 	if(..())
 		return
 
@@ -131,7 +136,7 @@
 			category = href_list["category"]
 
 	if(href_list["eject"] && !eject_lockout)
-		eject_materials_partial(href_list["eject"], href_list["mat_type"], text2num(href_list["amount"]))
+		eject_materials_partial(href_list["eject"], text2num(href_list["amount"]))
 
 	if(href_list["sync"])
 		sync()
@@ -288,7 +293,7 @@
 				hidden_mat = TRUE
 				continue
 		if(!hidden_mat)
-			. += list(list("mat" = capitalize(T), "amt" = materials[T], "type" = T?.type))
+			. += list(list("mat" = capitalize(T), "amt" = materials[T]))
 
 /obj/machinery/mecha_part_fabricator/proc/sync()
 	sync_message = "Error: no console found."
@@ -303,24 +308,21 @@
 		sync_message = "Sync complete."
 	update_categories()
 
-/obj/machinery/mecha_part_fabricator/proc/eject_materials_partial(var/material, var/mattype, var/amount)
-	material = lowertext(material)
-	var/obj/item/stack/material/S = new mattype(loc)
-	eject_lockout = TRUE
-	if(amount == 0)
-		amount = S.max_amount
-	if(amount == -1)
-		var/new_amount = input("Enter how many units of [material] to eject.", "Eject Material", 1) as num|null
-		if(!new_amount)
-			return
-		else
-			amount = min(new_amount, S.max_amount)
-	eject_lockout = FALSE
+/obj/machinery/mecha_part_fabricator/proc/eject_materials_partial(var/material, var/amount) // 0 amount = 0 means ejecting a full stack; -1 means eject everything
+	var/recursive = amount == -1 ? 1 : 0
+	var/matstring = lowertext(material)
+	var/datum/material/M = get_material_by_name(matstring)
+	if(recursive && materials[matstring] >= M.perunit)
+		eject_material_of_type(matstring)
+		return
 
-	world << "[materials[material]] | [S.perunit]"
-	var/ejected = min(round(materials[material] / S.perunit), amount)
+	var/obj/item/stack/material/S = M.place_sheet(get_turf(src))
+	if(amount <= 0)
+		amount = S.max_amount
+	var/ejected = min(round(materials[matstring] / S.perunit), amount)
 	S.amount = min(ejected, amount)
 	if(S.amount <= 0)
 		qdel(S)
 		return
-	materials[material] -= ejected * S.perunit
+	materials[matstring] -= ejected * S.perunit
+	update_busy()
